# chore: optimize maps.Values and maps.Keys