--- a/maps/maps.go
+++ b/maps/maps.go
@@ -42,22 +42,6 @@ func Map[K comparable, V any, K1 comparable, V1 any](m map[K]V, fn func(K, V) (K
 	return r
 }
 
-// Keys returns the keys of the map m.
-// The keys will be in an indeterminate order.
-func Keys[K comparable, V any](m map[K]V) []K {
-	if len(m) == 0 {
-		return nil
-	}
-
-	r := make([]K, 0, len(m))
-
-	for k := range m {
-		r = append(r, k)
-	}
-
-	return r
-}
-
 // KeysFunc applies the function fn to each key of the map m and returns a new slice with the results.
 // The keys will be in an indeterminate order.
 func KeysFunc[K comparable, V, R any](m map[K]V, fn func(K) R) []R {
@@ -69,18 +53,6 @@ func KeysFunc[K comparable, V, R any](m map[K]V, fn func(K) R) []R {
 
 	for k := range m {
 		r = append(r, fn(k))
-	}
-
-	return r
-}
-
-// Values returns the values of the map m.
-// The values will be in an indeterminate order.
-func Values[K comparable, V any](m map[K]V) []V {
-	r := make([]V, 0, len(m))
-
-	for _, v := range m {
-		r = append(r, v)
 	}
 
 	return r