--- a/xslices/xslices_test.go
+++ b/xslices/xslices_test.go
@@ -68,8 +68,6 @@ func TestFilterInPlace(t *testing.T) {
 	}
 
 	for name, tt := range tests {
-		tt := tt
-
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
 
@@ -138,8 +136,6 @@ func TestFilter(t *testing.T) {
 	}
 
 	for name, tt := range tests {
-		tt := tt
-
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()